to-do-list
I made this To Do List from the course online that I am learning full stack JS. This was my first time working with JS and I confess that it was really confusing for me. A lot of times I found my self dizzy and had no ideia what I was doing. It was really fun and frustating at the same time. Found a boring bug in the end and after ages searching, I saw that I missed a letter in a single word! I just put one letter and that it. hehe
It was really good for me to see that I need to pay attention in details.. I am looking forward to break my head more often with JS.